refactor(dashboard): rename files and clean up pages html code
re #187
