+++ apps/dashboard/src/routes.tsx
@@ -2,7 +2,7 @@ import "@fontsource/ibm-plex-sans"
 import "@rainbow-me/rainbowkit/styles.css"
 import { useContext, useMemo } from "react"
 import { createBrowserRouter, redirect, RouterProvider } from "react-router-dom"
-import { AuthContext } from "./context/authContext"
+import { AuthContext } from "./context/auth-context"
 import NotFoundPage from "./pages/404"
 import GroupPage from "./pages/group"
 import GroupsPage from "./pages/groups"
